+++ apps/builder/app/canvas/shared/scroll-new-instance-into-view.ts
@@ -7,9 +7,10 @@ import { $instances } from "~/shared/nano-states";
  */
 export const subscribeScrollNewInstanceIntoView = (
   debounceEffect: (callback: () => void) => void,
-  previousInstances: { current?: Instances }
+  previousInstances: { current?: Instances },
+  signal: AbortSignal
 ) => {
-  return $instances.subscribe((instances) => {
+  const unsubscribe = $instances.subscribe((instances) => {
     if (previousInstances.current === undefined) {
       previousInstances.current = instances;
       return;
@@ -36,4 +37,5 @@ export const subscribeScrollNewInstanceIntoView = (
       });
     });
   });
+  signal.addEventListener("abort", unsubscribe);
 };
