+++ apps/builder/app/shared/copy-paste/index.ts
@@ -1,1 +1,1 @@
-export { useCopyPaste } from "./use-copy-paste";
+export { initCopyPaste } from "./init-copy-paste";
